Fix RSS links in lists
Due to pagination, the relative "index.xml" urls were
broken as soon as "page/2/" was added to the URL.

The current fix is to take the first page from the pager,
as the first page is just "Posts/" for posts, "Blogvent" for blogvent, etc.

This is a bit of a hack, but seems to work just fine.